use seperate method for polling
